Reorganize arranger folder
