+++ lib/widgets/editors/arranger/events.dart
