+++ lib/widgets/editors/arranger/widgets/clip_sizer.dart
