+++ lib/widgets/editors/arranger/widgets/grid.dart
@@ -26,8 +26,8 @@ import 'package:anthem/widgets/editors/shared/helpers/grid_paint_helpers.dart';
 import 'package:anthem/widgets/editors/shared/helpers/types.dart';
 import 'package:flutter/widgets.dart';
 
-import 'arranger_view_model.dart';
-import 'helpers.dart';
+import '../arranger_view_model.dart';
+import '../helpers.dart';
 
 class ArrangerBackgroundPainter extends CustomPainterObserver {
   final double verticalScrollPosition;
