+++ lib/widgets/editors/arranger/widgets/track_header.dart
